+++ e2e/regression/cases/038_cli_json_field_addition/bqtail/schema.sql
@@ -1,0 +1,5 @@
+CREATE OR REPLACE TABLE dummy_v${parentIndex}_tmpl (
+      id         INT64 NOT NULL,
+      type_id    INT64 NOT NULL
+ );
+