+++ stage/load/request_test.go
@@ -73,7 +73,7 @@ func TestJob_NewLoadRequest(t *testing.T) {
 		},
 	}
 
-	for _, useCase := range useCases[8:] {
+	for _, useCase := range useCases {
 		ctx := context.Background()
 		err := loadTestAsset(ctx, &useCase.process, path.Join(useCase.caseURL, "process.json"))
 		if !assert.Nil(t, err, useCase.description) {