+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## March 30 2020 2.1.0
  * Added DMLAppend option
+ * Added AllowFieldAddition that works also with JSON/CSV source format
 
 ## March 27  2020 2.0.3
  * Refactored retries